Tersoff model implementation   (#175)
* Add Tersoff potential model

- The first draft of the Tersoff potential model is implemented.
  The model still has problems to compute energy correctly.
  It will be updated.

* Initial commit

Tersoff model, one of the bond-order model, is being implemented.
This is the first trial, so it has a lot of problems.

* Add Tersoff model

Tersoff model was implemented. Neighbor list should be provided.
The current model can be applied to the unary system.

* Update parameter reading function in Tersoff model

Parameter reading function 'load_lammps_tersoff_parameters' is added.
This function can read a parameter file, which is written as LAMMPS
format. The function returns parameter list that can be used in the
Tersoff model. In order to use parameter list, Tersoff model and its
test are updated.

Miscellaneous changes are listed below:
  - Comments for functions are updated.
  - An example of the parameter file 'Si.tersoff' is added to the
'tests/data' directory for testing.'.

* Fix minor problems

Some minor problems are fixed.
  - A comment in tersoff parameter function is longer than 80 character.
    So, the comment is divided into two lines.
  - A test for SW model is reverted as the original version.

* Correcting for comments