Update simulate.py (#179)
* Update simulate.py

* Update simulate.py

Added options to pass in key word arguments to the update_sigma_fn in Brownian Dynamics simulation

* Update simulate.py

* Update simulate.py

* Update documentation of mass and gamma

In response to @cpgoodri 's comment. Rolling back to the original definition to ensure backward compatibility, and updated the documentation to clarify the definition of gamma. Also, added the option for user to pass in gamma in the apply_fn similarly to how temperature can be passed in.

* Update quantity.py

* Update simulate.py

---------

Co-authored-by: Abhijeet Gangan <abhijeetgangan@g.ucla.edu>